requirements.txt
django<2.1
psycopg2<2.8
django-markdownify==0.3.0
django-crispy-forms==1.7.0
factory_boy<3.0
python-dotenv
celery<4.2
celery[redis]
django-celery-results<2.0
redis<2.11
djangorestframework<3.8
